server/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
bidict==0.23.1
boto3==1.28.3
botocore==1.31.85
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.1.7
django-admin-json-editor==0.2.3
django-cors-headers==4.7.0
django-filter==25.1
django-json-widget==2.0.1
django-unfold==0.52.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
fastapi==0.109.0
ffmpeg-python==0.2.0
future==1.0.0
h11==0.14.0
idna==3.7
inflection==0.5.1
jmespath==1.0.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.9
pydantic==1.10.13
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.9.1
python-socketio==5.11.0
pytz==2025.1
PyYAML==6.0.2
requests==2.31.0
s3transfer==0.6.2
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.35.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.27.0
wsproto==1.2.0
